+++ packages/sdk-components-react/src/components.ts
@@ -30,4 +30,4 @@ export { Vimeo } from "./vimeo";
 export { VimeoPreviewImage } from "./vimeo-preview-image";
 export { VimeoPlayButton } from "./vimeo-play-button";
 export { VimeoSpinner } from "./vimeo-spinner";
-export { XmlElement } from "./xml-element";
+export { XmlNode } from "./xml-node";
