+++ packages/sdk-components-react/src/metas.ts
@@ -30,4 +30,4 @@ export { meta as Vimeo } from "./vimeo.ws";
 export { meta as VimeoPreviewImage } from "./vimeo-preview-image.ws";
 export { meta as VimeoPlayButton } from "./vimeo-play-button.ws";
 export { meta as VimeoSpinner } from "./vimeo-spinner.ws";
-export { meta as XmlElement } from "./xml-element.ws";
+export { meta as XmlNode } from "./xml-node.ws";
