+++ packages/sdk-components-react/src/props.ts
@@ -30,4 +30,4 @@ export { propsMeta as Vimeo } from "./vimeo.ws";
 export { propsMeta as VimeoPreviewImage } from "./vimeo-preview-image.ws";
 export { propsMeta as VimeoPlayButton } from "./vimeo-play-button.ws";
 export { propsMeta as VimeoSpinner } from "./vimeo-spinner.ws";
-export { propsMeta as XmlElement } from "./xml-element.ws";
+export { propsMeta as XmlNode } from "./xml-node.ws";
